+++ test/ensure_build_test.dart
@@ -5,11 +5,11 @@ import 'package:build_verify/build_verify.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('ensure_build', () {
+  test('ensure_build', () async {
     // TODO(sigurdm): investigate why the build doesn't come clean on
     // Windows.
     if (!Platform.isWindows) {
-      expectBuildClean();
+      await expectBuildClean();
     }
   });
 }